requirements.txt
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
cffi==1.13.2
Click==7.0
constantly==15.1.0
cryptography==2.8
hkdf==0.0.3
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
magic-wormhole==0.11.2
packaging==19.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.5
PyQt5==5.13.0
PyQt5-sip==4.19.18
qt5reactor==0.5
service-identity==18.1.0
six==1.12.0
spake2==0.8
tqdm==4.36.1
Twisted==19.7.0
txaio==18.8.1
txtorcon==19.1.0
zope.event==4.4
zope.interface==4.6.0
